Complete Application code

+++ App/app/android/app/src/main/java/com/example/app/UserRssi.java
@@ -9,8 +9,10 @@ public class UserRssi implements Comparator<MinewBeacon> {
 
     @Override
     public int compare(MinewBeacon minewBeacon, MinewBeacon t1) {
-        float floatValue1 = minewBeacon.getBeaconValue(BeaconValueIndex.MinewBeaconValueIndex_RSSI).getFloatValue();
-        float floatValue2 = t1.getBeaconValue(BeaconValueIndex.MinewBeaconValueIndex_RSSI).getFloatValue();
+        // float floatValue1 = minewBeacon.getBeaconValue(BeaconValueIndex.MinewBeaconValueIndex_RSSI).getFloatValue();
+        // float floatValue2 = t1.getBeaconValue(BeaconValueIndex.MinewBeaconValueIndex_RSSI).getFloatValue();
+        int floatValue1 = minewBeacon.getBeaconValue(BeaconValueIndex.MinewBeaconValueIndex_RSSI).getIntValue();
+        int floatValue2 = t1.getBeaconValue(BeaconValueIndex.MinewBeaconValueIndex_RSSI).getIntValue();
         if(floatValue1<floatValue2){
             return 1;
         }else if(floatValue1==floatValue2){
